requirements.txt
sentence-transformers
qdrant-client
transformers
pillow
streamlit
tqdm